requirements.txt
pyyaml
pydantic
pydantic[email]
